# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.badgeResponse:
    properties:
      description:
        type: string
      earned_at:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  handlers.interestResponse:
    properties:
      category:
        type: string
      genre_tag_id:
        type: integer
      icon:
        type: string
      name:
        type: string
    type: object
...
  handlers.proficiencyResponse:
    properties:
      category:
        type: string
      genre_name:
        type: string
      genre_tag_id:
        type: integer
      icon:
        type: string
      level:
        type: integer
      xp:
        type: integer
    type: object
...
  handlers.updateInterestsRequest:
    properties:
      genre_tag_ids:
        items:
          type: integer
        type: array
    required:
    - genre_tag_ids
    type: object
  handlers.updateInterestsResponse:
    properties:
      interests:
        items:
          $ref: '#/definitions/handlers.interestResponse'
        type: array
      reload_count_remaining:
        type: integer
    type: object
  handlers.updateMeRequest:
    properties:
      display_name:
        type: string
      search_radius:
        type: integer
    type: object
  handlers.updateMeResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      level:
        type: integer
      reload_count_remaining:
        type: integer
      search_radius:
        type: integer
      settings_json:
        type: string
      streak_count:
        type: integer
      streak_last:
        type: string
      total_xp:
        type: integer
      updated_at:
        type: string
    type: object
...
  handlers.userStatsResponse:
    properties:
      breakout_visits:
        type: integer
      challenge_visits:
        type: integer
      level:
        type: integer
      streak_count:
        type: integer
      streak_last:
        type: string
      total_visits:
        type: integer
      total_xp:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      level:
        type: integer
      search_radius:
        type: integer
      settings_json:
        type: string
      streak_count:
        type: integer
      streak_last:
        type: string
      total_xp:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: RoambleのバックエンドAPIドキュメント
  title: Roamble API
  version: "1.0"
paths:
...
  /api/users/me:
    delete:
      description: JWT認証済みユーザーのアカウントと関連データを物理削除する
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: アカウント削除
      tags:
      - Users
    get:
      description: JWT認証済みユーザーの情報を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー情報取得
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: JWT認証済みユーザーの表示名・提案半径を更新する（各フィールドはオプショナル）。refresh_suggestions=true
        かつ search_radius 変更時はリロード1回分を消費して提案キャッシュをクリアする。
      parameters:
      - description: true の場合（提案半径変更時）、提案キャッシュをクリアしリロードカウントを消費する
        in: query
        name: refresh_suggestions
        type: boolean
      - description: 更新情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.updateMeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー情報更新
      tags:
      - Users
  /api/users/me/badges:
    get:
      description: JWT認証済みユーザーの獲得バッジ一覧を返す（獲得日時の降順）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.badgeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 獲得バッジ一覧取得
      tags:
      - Users
  /api/users/me/interests:
    get:
      description: JWT認証済みユーザーの興味タグ一覧を返す（ジャンル名昇順）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.interestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー興味タグ取得
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: JWT認証済みユーザーの興味タグを一括更新する（3つ以上必須）。refresh_suggestions=true を指定するとリロード1回分を消費して提案キャッシュをクリアする。
      parameters:
      - description: true の場合、提案キャッシュをクリアしリロードカウントを消費する
        in: query
        name: refresh_suggestions
        type: boolean
      - description: 興味タグIDリスト
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateInterestsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.updateInterestsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー興味タグ更新
      tags:
      - Users
  /api/users/me/proficiency:
    get:
      description: JWT認証済みユーザーのジャンル別熟練度一覧を返す（XP降順）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.proficiencyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ジャンル別熟練度取得
      tags:
      - Users
  /api/users/me/stats:
    get:
      description: JWT認証済みユーザーのレベル・XP・ストリーク・訪問統計を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userStatsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー統計情報取得
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: JWTアクセストークン（Bearer形式）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"